AtomicMap Base
